Move core routing to subdirectory
